+++ Tests/Fluxor.Blazor.Web.UnitTests/Components/FluxorLayoutTests/Dispose.cs
@@ -3,28 +3,29 @@ using Xunit;
 
 namespace Fluxor.Blazor.Web.UnitTests.Components.FluxorComponentTests
 {
-	public partial class FluxorComponentTests
+	public partial class FluxorLayoutTests
 	{
-		public class OnInitialized
+		public class Dispose
 		{
-			private readonly FluxorComponentWithStateProperties Subject;
+			private readonly FluxorLayoutWithStateProperties Subject;
 			private readonly MockState<int> MockState1;
 			private readonly MockState<int> MockState2;
 
 			[Fact]
-			public void SubscribesToStateProperties()
+			public void UnsubscribesFromStateProperties()
 			{
 				Subject.ExecuteOnInitialized();
+				Subject.Dispose();
 
-				Assert.Equal(1, MockState1.SubscribeCount);
-				Assert.Equal(1, MockState2.SubscribeCount);
+				Assert.Equal(1, MockState1.UnsubscribeCount);
+				Assert.Equal(1, MockState2.UnsubscribeCount);
 			}
 
-			public OnInitialized()
+			public Dispose()
 			{
 				MockState1 = new MockState<int>();
 				MockState2 = new MockState<int>();
-				Subject = new FluxorComponentWithStateProperties {
+				Subject = new FluxorLayoutWithStateProperties {
 					State1 = MockState1,
 					State2 = MockState2
 				};