+++ Tests/Fluxor.Blazor.Web.UnitTests/Components/FluxorLayoutTests/OnInitialized.cs
@@ -3,11 +3,11 @@ using Xunit;
 
 namespace Fluxor.Blazor.Web.UnitTests.Components.FluxorComponentTests
 {
-	public partial class FluxorComponentTests
+	public partial class FluxorLayoutTests
 	{
 		public class OnInitialized
 		{
-			private readonly FluxorComponentWithStateProperties Subject;
+			private readonly FluxorLayoutWithStateProperties Subject;
 			private readonly MockState<int> MockState1;
 			private readonly MockState<int> MockState2;
 
@@ -24,7 +24,7 @@ namespace Fluxor.Blazor.Web.UnitTests.Components.FluxorComponentTests
 			{
 				MockState1 = new MockState<int>();
 				MockState2 = new MockState<int>();
-				Subject = new FluxorComponentWithStateProperties {
+				Subject = new FluxorLayoutWithStateProperties {
 					State1 = MockState1,
 					State2 = MockState2
 				};